don't delete tables if they don't exist

+++ models.py
@@ -1,8 +1,7 @@
 import os
 
-import psycopg2
 from flask_sqlalchemy import SQLAlchemy
-from sqlalchemy import Column, String, Integer, create_engine
+from sqlalchemy import Column, String, Integer, inspect
 from sqlalchemy.ext.compiler import compiles
 from sqlalchemy.schema import DropTable
 
@@ -42,12 +41,11 @@ def db_drop_and_create_all():
         drops the database tables and starts fresh
         can be used to initialize a clean database
     """
-    try:
+    inspector_gadget = inspect(db.engine)
+    if inspector_gadget.has_table("player"):
         db.session.query(Player).delete()
         db.session.query(Tournament).delete()
-        db.session.commit()
-    except psycopg2.ProgrammingError:
-        pass
+    db.session.commit()
     db.create_all()
